backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
psutil==5.9.6
scapy==2.5.0
python-dotenv==1.0.0
